+++ package.json
@@ -18,6 +18,9 @@
     "start": "next start",
     "ragequit": "rm -rf .next && npm run db:push && npm run dev"
   },
+  "prisma": {
+    "seed": "tsx prisma/seeds/index.ts"
+  },
   "dependencies": {
     "@auth/prisma-adapter": "^1.4.0",
     "@hookform/resolvers": "^3.3.4",
@@ -30,6 +33,7 @@
     "@radix-ui/react-dropdown-menu": "^2.0.6",
     "@radix-ui/react-icons": "^1.3.0",
     "@radix-ui/react-label": "^2.0.2",
+    "@radix-ui/react-scroll-area": "^1.0.5",
     "@radix-ui/react-select": "^2.0.0",
     "@radix-ui/react-slot": "^1.0.2",
     "@radix-ui/react-toast": "^1.1.5",
@@ -57,7 +61,7 @@
   "devDependencies": {
     "@types/eslint": "^8.56.2",
     "@types/lodash": "^4.17.0",
-    "@types/node": "^20.11.20",
+    "@types/node": "^20.12.7",
     "@types/nodemailer": "^6.4.14",
     "@types/react": "^18.2.57",
     "@types/react-dom": "^18.2.19",
@@ -73,6 +77,7 @@
     "prisma": "^5.10.2",
     "tailwindcss": "^3.4.1",
     "tailwindcss-animate": "^1.0.7",
+    "tsx": "^4.7.2",
     "typescript": "^5.4.2"
   },
   "ct3aMetadata": {
