+++ prisma/schema.prisma
@@ -67,7 +67,7 @@ model User {
   email            String?      @unique
   emailVerified    DateTime?
   image            String?
-  stripeCustomerId String?
+  stripeCustomerId String?      @unique
   accounts         Account[]
   sessions         Session[]
   memberships      Membership[]
