+++ src/components/molecules/membershipTemplateCard.tsx
@@ -1,4 +1,5 @@
 import { Card, CardContent } from "@/components/ui/card";
+import {PricePeriod, PriceUnit} from "@prisma/client";
 
 interface Price {
   value: number;
@@ -6,31 +7,21 @@ interface Price {
   unit: PriceUnit;
 }
 
-export enum PriceUnit {
-  Eur = "EUR",
-  Usd = "USD",
-}
-
-export enum PricePeriod {
-  Monthly = "month",
-  Yearly = "year",
-}
-
-interface Membership {
+interface MembershipTemplateCardProps {
   showPrice?: boolean;
   title: string;
   features: string[];
-  description?: string;
+  description: string | null;
   price: Price;
 }
 
-export function MembershipCard({
+export function MembershipTemplateCard({
   title,
   features,
   description,
   price,
   showPrice = true,
-}: Membership) {
+}: MembershipTemplateCardProps) {
   const moneyFormatter = new Intl.NumberFormat("en-US", {
     currency: price.unit,
     minimumFractionDigits: 0,
@@ -38,7 +29,7 @@ export function MembershipCard({
   });
 
   return (
-    <Card>
+    <Card className={'w-[320px]'}>
       <CardContent className={"p-6"}>
         <div className="mb-4 flex items-start justify-between">
           <h4 className={"text-lg font-semibold"}>{title}</h4>
